updated project

+++ Django with mongodb/core/vege/views.py
@@ -66,7 +66,11 @@ def update_receipe(request, id):
     return render(request, 'update_receipes.html', context)
 
 def delete_receipe(request, id):
-    id = int(id)  # Convert to integer
-    receipe = get_object_or_404(Receipe, id=id)
-    receipe.delete()
-    return redirect('/receipes/')
+    if id is not None:
+        id = int(id)
+        receipe = get_object_or_404(Receipe, id=id)
+        receipe.delete()
+        return redirect('/receipes/')
+    else:
+        # Handle the case where 'id' is None (or redirect to an error page)
+        return HttpResponse("Invalid request")
